requirements.txt
requests==2.28.1
google-genai
dotenv